reorganize and add more tests
