+++ frontend/src/metabase/sharing/components/SharingSidebar/SharingSidebar.jsx
