+++ frontend/src/metabase/sharing/components/SharingSidebar/index.jsx
@@ -1,0 +1,1 @@
+export * from "./SharingSidebar";
