+++ frontend/src/metabase/sharing/components/SharingSidebar/tests/premium.unit.spec.js
@@ -1,0 +1,34 @@
+import userEvent from "@testing-library/user-event";
+import { screen } from "__support__/ui";
+
+import { setup, hasAdvancedFilterOptions } from "./setup";
+
+describe("SharingSidebar Premium Features", () => {
+  const tokenFeatures = {
+    advanced_config: true,
+  };
+
+  describe("Email Subscription sidebar", () => {
+    it("should show advanced filtering options with the correct feature flag", async () => {
+      setup({ email: true, tokenFeatures, hasEnterprisePlugins: true });
+
+      userEvent.click(await screen.findByText("Email it"));
+
+      await screen.findByText("Email this dashboard");
+
+      expect(hasAdvancedFilterOptions(screen)).toBe(true);
+    });
+  });
+
+  describe("Slack Subscription sidebar", () => {
+    it("should show advanced filtering options with the correct feature flag", async () => {
+      setup({ slack: true, tokenFeatures, hasEnterprisePlugins: true });
+
+      userEvent.click(await screen.findByText("Send it to Slack"));
+
+      await screen.findByText("Send this dashboard to Slack");
+
+      expect(hasAdvancedFilterOptions(screen)).toBe(true);
+    });
+  });
+});
